TaxEstimator_Final_Project/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
openai==1.59.7
pymongo==4.10.1
azure-storage-blob==12.24.0

